requirements.txt
pydantic
anthropic
openai
langchain-text-splitters
python-dotenv
psycopg2-binary
pgvector
sqlalchemy
python-fasthtml
fastapi
requests
uvicorn

# # to enable use in jupyter lab 
# ipykernel


